review_spot_backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pytz==2024.1
PyYAML==6.0.1
setuptools==72.1.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
